+++ src/main/java/org/truffleruby/core/thread/ThreadManager.java
@@ -56,6 +56,7 @@ import org.truffleruby.language.control.DynamicReturnException;
 import org.truffleruby.language.control.ExitException;
 import org.truffleruby.language.control.KillException;
 import org.truffleruby.language.control.RaiseException;
+import org.truffleruby.language.dispatch.DispatchNode;
 import org.truffleruby.language.objects.shared.SharedObjects;
 
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
@@ -375,14 +376,10 @@ public class ThreadManager {
 
             if (!isSystemExit && thread.reportOnException) {
                 final TruffleSafepoint safepoint = TruffleSafepoint.getCurrent();
-
                 boolean sideEffects = safepoint.setAllowSideEffects(false);
                 try {
-                    RubyContext.send(
-                            context.getCoreLibrary().truffleThreadOperationsModule,
-                            "report_exception",
-                            thread,
-                            exception);
+                    DispatchNode.getUncached().call(context.getCoreLibrary().truffleThreadOperationsModule,
+                            "report_exception", thread, exception);
                 } finally {
                     safepoint.setAllowSideEffects(sideEffects);
                 }