+++ src/main/java/org/truffleruby/language/EmitWarningsNode.java
@@ -17,6 +17,7 @@ import org.truffleruby.core.rope.Rope;
 import org.truffleruby.core.string.RubyString;
 import org.truffleruby.core.string.StringOperations;
 import org.truffleruby.language.control.RaiseException;
+import org.truffleruby.language.dispatch.DispatchNode;
 import org.truffleruby.parser.RubyDeferredWarnings;
 
 import java.io.IOException;
@@ -70,7 +71,7 @@ public class EmitWarningsNode extends RubyContextSourceNode {
             final Rope messageRope = StringOperations.encodeRope(message, UTF8Encoding.INSTANCE);
             final RubyString messageString = StringOperations
                     .createUTF8String(context, context.getLanguageSlow(), messageRope);
-            RubyContext.send(warning, "warn", messageString);
+            DispatchNode.getUncached().call(warning, "warn", messageString);
         } else {
             try {
                 context.getEnvErrStream().write(message.getBytes(StandardCharsets.UTF_8));