+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Compatibility:
 * Fix `Float#/` when dividing by `Rational` (@bjfish).
 * `Process.euid=` should accept String (#2615, @ngtban).
 * Fix `instance_variable_get` and `instance_variable_set` for immutable objects (@bjfish).
+* `Thread#raise(exc, message)` now calls `exc.exception` in the target thread like CRuby (@eregon).
 
 Performance: